% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifying_sfbdrt.R
\name{modified_sfbdrt}
\alias{modified_sfbdrt}
\title{modified_sfbdrt: Modified the sfbdrt(sim.fbd.rateshift.taxa) function
 to return fossil and SAtree objects}
\usage{
modified_sfbdrt(n, numbsim, lambda, mu, psi, times, complete = FALSE)
}
\arguments{
\item{n}{Number of extant taxa}

\item{numbsim}{Number of trees to simulate}

\item{lambda}{Speciation rate}

\item{mu}{Extinction rate}

\item{psi}{Fossil sampling rate. Input as a vector.}

\item{times}{Vector of time intervals.}

\item{complete}{Defaults to FALSE. If TRUE, the tree including the extint
lineages and non-sampled lineages is returned.}
}
\value{
list(f,trees). Returns fossil object, f, and simulated SAtrees with n extant
sampled tips
}
\description{
sim.fbd.rateshift.taxa function from the FossilSim Package,
but it returns a list of the fossil along with the SAtree object.
It also adds an (h) column to the fossil dataframe so that the fossil count
functions (summary statistic) will work.
}
