% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifying_sfbdrt.R
\name{n.ages}
\alias{n.ages}
\title{n.ages: Accompanying function for sim.fbd.rateshift.taxa from FossilSim package}
\usage{
n.ages(tree)
}
\arguments{
\item{tree}{}
}
\value{
node.ages
}
\description{
n.ages: Accompanying function for sim.fbd.rateshift.taxa from FossilSim package
}
