% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_simulations.R
\name{sim_multiple_rows}
\alias{sim_multiple_rows}
\title{sim_multiple_rows: Simulate tree from multiple rows and apply summary statistics}
\usage{
sim_multiple_rows(
  data,
  extant,
  mus,
  lambdas,
  psis,
  intervals,
  by = FALSE,
  path = c()
)
}
\arguments{
\item{data}{data frame}

\item{extant}{Number of extant taxa}

\item{mus}{Extinction rate}

\item{lambdas}{Speciation rate}

\item{psis}{Fossil sampling rate. Input as a vector}

\item{intervals}{Vector of time intervals or time bins}

\item{by}{Selects for specific rows. For example, by = 100 refers to sampling
every 100th row. Defaults to FALSE. If FALSE, every row is sampled to produce
a simulated phylo object.}

\item{path}{Specify path to save}
}
\value{
summlist. An object describing useful summary statistics like
total fossil count, fossil count per interval, and origin time
}
\description{
Function runs multiple simulations. Can be specified by row.
Also provides summary statistics like total fossil count, fossil
count per interval, and origin time.
}
