% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sfbdrt.R
\name{sim_single_row}
\alias{sim_single_row}
\title{sim_single_row: Simulating trees from a single row of the posterior trace}
\usage{
sim_single_row(n_extant, n_trees, df, mu, lambda, psi, row, mers)
}
\arguments{
\item{n_extant}{Number of extant taxa}

\item{n_trees}{Number of trees to simulate}

\item{df}{data frame}

\item{mu}{Extinction rate}

\item{lambda}{Speciation rate}

\item{psi}{Fossil sampling rate. Input as a vector}

\item{row}{Row to sample}

\item{mers}{Vector of time intervals or time bins}
}
\value{
modelledfoss, the simulated list of phylo objects
with parameter specifications
}
\description{
Runs modified_sfbdrt with parameter specifications to simulate
tree from a single, specified row from the phylogenetic posterior prediction.
}
