% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{summarize_tree}
\alias{summarize_tree}
\title{summarize_tree: Summarizes simulated tree}
\usage{
summarize_tree(phylolist, interval.ages)
}
\arguments{
\item{phylolist}{simulated list of phylo objects}

\item{interval.ages}{Vector of time intervals}
}
\value{
A vector of origin time, total fossils count, and fossil count per interval
}
\description{
summarize_tree calculates useful statistics like
the origin time, total fossils, and fossil count per time interval for
a particular simulation.
}
